add condition

+++ backend/src/main/java/io/spring/api/CommentsApi.java
@@ -76,6 +76,9 @@ public class CommentsApi {
         .findById(item.getId(), commentId)
         .map(
             comment -> {
+              if (!AuthorizationService.canWriteComment(user, item, comment)) {
+                throw new NoAuthorizationException();
+              }
               commentRepository.remove(comment);
               return ResponseEntity.noContent().build();
             })
